Added System.DoubleNumerics to the core project Converted Transform.cs to use System.DoubleNumerics Added a new TransformTests.cs file to test the new transform class Changed Converters to use System.DoubleNumerics

+++ Objects/Converters/ConverterNavisworks/ConverterNavisworks/ConverterNavisworks.Geometry.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
-using System.Numerics;
+using System.DoubleNumerics;
 using Autodesk.Navisworks.Api;
 using Autodesk.Navisworks.Api.Interop.ComApi;
 using Objects.Geometry;
@@ -213,7 +213,7 @@ public class PrimitiveProcessor : InwSimplePrimitivesCB
   private static Vector3 VectorFromVertex(InwSimpleVertex v)
   {
     var arrayV = (Array)v.coord;
-    return new Vector3((float)arrayV.GetValue(1), (float)arrayV.GetValue(2), (float)arrayV.GetValue(3));
+    return new Vector3(arrayV.GetValue(1), arrayV.GetValue(2), arrayV.GetValue(3));
   }
 }
