+++ Objects/Converters/ConverterRevit/ConverterRevitShared/Partial Classes/ConvertGeometry.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Numerics;
+using System.DoubleNumerics;
 using Autodesk.Revit.DB;
 using Autodesk.Revit.DB.PointClouds;
 using Objects.Converters.DxfConverter;
@@ -658,9 +658,9 @@ namespace Objects.Converter.Revit
         if (points.Count != 4) return false;
 
         var matrix = new Matrix4x4(
-          (float)points[0].X, (float)points[1].X, (float)points[2].X, (float)points[3].X,
-          (float)points[0].Y, (float)points[1].Y, (float)points[2].Y, (float)points[3].Y,
-          (float)points[0].Z, (float)points[1].Z, (float)points[2].Z, (float)points[3].Z,
+          points[0].X, points[1].X, points[2].X, points[3].X,
+          points[0].Y, points[1].Y, points[2].Y, points[3].Y,
+          points[0].Z, points[1].Z, points[2].Z, points[3].Z,
           1, 1, 1, 1
         );
         return matrix.GetDeterminant() != 0;
